+++ Notechigima/src/main/java/com/capstone/notechigima/controller/AdviceController.java
@@ -3,7 +3,7 @@ package com.capstone.notechigima.controller;
 import com.capstone.notechigima.config.BaseException;
 import com.capstone.notechigima.config.BaseResponse;
 import com.capstone.notechigima.config.BaseResponseStatus;
-import com.capstone.notechigima.model.comment.PostCommentRequestDTO;
+import com.capstone.notechigima.dto.comment.PostCommentRequestDTO;
 import com.capstone.notechigima.service.AdviceService;
 import com.capstone.notechigima.service.CommentService;
 import io.swagger.v3.oas.annotations.Operation;
