+++ Notechigima/src/main/java/com/capstone/notechigima/controller/SubjectController.java
@@ -2,7 +2,7 @@ package com.capstone.notechigima.controller;
 
 import com.capstone.notechigima.config.BaseResponse;
 import com.capstone.notechigima.config.BaseResponseStatus;
-import com.capstone.notechigima.model.topic.TopicResponseDTO;
+import com.capstone.notechigima.dto.topic.TopicResponseDTO;
 import com.capstone.notechigima.service.TopicService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
