+++ Notechigima/src/main/java/com/capstone/notechigima/controller/TopicController.java
@@ -3,9 +3,9 @@ package com.capstone.notechigima.controller;
 import com.capstone.notechigima.config.BaseException;
 import com.capstone.notechigima.config.BaseResponseStatus;
 import com.capstone.notechigima.config.BaseResponse;
-import com.capstone.notechigima.model.advice.AdviceResponseDTO;
-import com.capstone.notechigima.model.note.GetNoteSummarizedDTO;
-import com.capstone.notechigima.model.note.PostNoteRequestDTO;
+import com.capstone.notechigima.dto.advice.AdviceResponseDTO;
+import com.capstone.notechigima.dto.note.GetNoteSummarizedDTO;
+import com.capstone.notechigima.dto.note.PostNoteRequestDTO;
 import com.capstone.notechigima.service.AdviceService;
 import com.capstone.notechigima.service.NoteService;
 import com.capstone.notechigima.service.TopicService;
