+++ Notechigima/src/main/java/com/capstone/notechigima/controller/UserController.java
@@ -3,7 +3,7 @@ package com.capstone.notechigima.controller;
 import com.capstone.notechigima.config.BaseException;
 import com.capstone.notechigima.config.BaseResponse;
 import com.capstone.notechigima.config.BaseResponseStatus;
-import com.capstone.notechigima.model.group.GetGroupResponseDTO;
+import com.capstone.notechigima.dto.group.GetGroupResponseDTO;
 import com.capstone.notechigima.service.GroupService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
