+++ Notechigima/src/main/java/com/capstone/notechigima/dto/comment/PostCommentRequestDTO.java
@@ -1,4 +1,4 @@
-package com.capstone.notechigima.model.comment;
+package com.capstone.notechigima.dto.comment;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
