+++ Notechigima/src/main/java/com/capstone/notechigima/dto/group/PostGroupRequestDTO.java
@@ -1,4 +1,4 @@
-package com.capstone.notechigima.model.group;
+package com.capstone.notechigima.dto.group;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
