+++ Notechigima/src/main/java/com/capstone/notechigima/service/AdviceService.java
@@ -1,6 +1,6 @@
 package com.capstone.notechigima.service;
 
-import com.capstone.notechigima.model.advice.AdviceResponseDTO;
+import com.capstone.notechigima.dto.advice.AdviceResponseDTO;
 
 import java.util.List;
