+++ Notechigima/src/main/java/com/capstone/notechigima/service/GroupServiceImpl.java
@@ -1,10 +1,10 @@
 package com.capstone.notechigima.service;
 
-import com.capstone.notechigima.model.ModelMapper;
+import com.capstone.notechigima.dto.ModelMapper;
 import com.capstone.notechigima.domain.group.GroupCreateEntity;
 import com.capstone.notechigima.domain.users.MemberEntity;
-import com.capstone.notechigima.model.group.GetGroupResponseDTO;
-import com.capstone.notechigima.model.group.PostGroupRequestDTO;
+import com.capstone.notechigima.dto.group.GetGroupResponseDTO;
+import com.capstone.notechigima.dto.group.PostGroupRequestDTO;
 import com.capstone.notechigima.repository.GroupRepository;
 import org.springframework.stereotype.Service;
