+++ Notechigima/src/main/java/com/capstone/notechigima/service/NoteServiceImpl.java
@@ -1,14 +1,14 @@
 package com.capstone.notechigima.service;
 
 import com.capstone.notechigima.config.BaseException;
-import com.capstone.notechigima.model.ModelMapper;
+import com.capstone.notechigima.dto.ModelMapper;
 import com.capstone.notechigima.domain.note.NoteDetailEntity;
 import com.capstone.notechigima.domain.sentence.SentenceEntity;
 import com.capstone.notechigima.domain.note.NoteInsertEntity;
-import com.capstone.notechigima.model.note.GetNoteResponseDTO;
-import com.capstone.notechigima.model.note.GetNoteSummarizedDTO;
-import com.capstone.notechigima.model.note.PostNoteRequestDTO;
-import com.capstone.notechigima.model.sentence.SentenceResponseDTO;
+import com.capstone.notechigima.dto.note.GetNoteResponseDTO;
+import com.capstone.notechigima.dto.note.GetNoteSummarizedDTO;
+import com.capstone.notechigima.dto.note.PostNoteRequestDTO;
+import com.capstone.notechigima.dto.sentence.SentenceResponseDTO;
 import com.capstone.notechigima.repository.NoteRepository;
 import com.capstone.notechigima.repository.SentenceRepository;
 import org.springframework.stereotype.Service;
