+++ Notechigima/src/main/java/com/capstone/notechigima/service/TopicServiceImpl.java
@@ -1,11 +1,11 @@
 package com.capstone.notechigima.service;
 
-import com.capstone.notechigima.model.ModelMapper;
+import com.capstone.notechigima.dto.ModelMapper;
 import com.capstone.notechigima.domain.advice.AdviceEntity;
 import com.capstone.notechigima.domain.advice.AdviceType;
 import com.capstone.notechigima.domain.sentence.SentenceWithWriterEntity;
-import com.capstone.notechigima.model.advice.AdviceInferenceRequestVO;
-import com.capstone.notechigima.model.topic.TopicResponseDTO;
+import com.capstone.notechigima.dto.advice.AdviceInferenceRequestVO;
+import com.capstone.notechigima.dto.topic.TopicResponseDTO;
 import com.capstone.notechigima.repository.AdviceRepository;
 import com.capstone.notechigima.repository.NoteRepository;
 import com.capstone.notechigima.repository.TopicRepository;
