+++ Notechigima/src/main/java/com/capstone/notechigima/service/UserServiceImpl.java
@@ -1,7 +1,7 @@
 package com.capstone.notechigima.service;
 
-import com.capstone.notechigima.model.ModelMapper;
-import com.capstone.notechigima.model.users.GetUserResponseDTO;
+import com.capstone.notechigima.dto.ModelMapper;
+import com.capstone.notechigima.dto.users.GetUserResponseDTO;
 import com.capstone.notechigima.repository.UserRepository;
 import org.springframework.stereotype.Service;
